requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework>=3.15,<4
django-cors-headers>=4.7,<5
python-dotenv>=1.0,<2
psycopg[binary]>=3.2,<4
sqlparse==0.5.5
tzdata==2026.1
anthropic
